requirements.txt
pandas>=2.0.0
pandasai>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
python-dotenv>=1.0.0
openpyxl>=3.1.0  # 用于Excel文件支持
pyarrow>=14.0.0  # 用于parquet和feather文件支持 